Working webpage
there is an issue with the db calculation with time worked, it gives negative amounts of time if clocked in then out quickly..... perhaps we need a clock out Page? when someone clocks in our out it sends them to a dead end page that logs ppl out?

also hitting back and forward buttons step with button presses for some reason? it does not seem to resend db querys though.